fix(core): use PlayerSlot(byte) instead of PlayerSlot(int) for slot lookups
PlayerSlot(int) routes through the EntityIndex constructor which
produces a different value than the original slot. Using (byte) cast
matches PlayerSlot.AsPrimitive() and correctly resolves the client.

+++ MapChooserSharpMS/Modules/RockTheVote/Services/RtvService.cs
@@ -90,7 +90,7 @@ internal sealed class RtvService(
 
     public RtvExecutionResult AddClientToRtv(int slot)
     {
-        var client = plugin.SharedSystem.GetClientManager().GetGameClient(new PlayerSlot(slot));
+        var client = plugin.SharedSystem.GetClientManager().GetGameClient(new PlayerSlot((byte)slot));
 
         if (client == null)
             return RtvExecutionResult.NotAllowed;
